move log config into log package
